internal: Reorganize files
